requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0

# Financial data and APIs
yfinance>=0.2.18
polygon-api-client>=1.12.0
alpaca-trade-api>=3.0.0

# Web scraping
beautifulsoup4>=4.12.0
selenium>=4.15.0
lxml>=4.9.0

# Technical analysis
ta-lib>=0.4.25
pandas-ta>=0.3.14b

# Machine learning and NLP
scikit-learn>=1.3.0
textblob>=0.17.1
vaderSentiment>=3.3.2

# Database
sqlite3
sqlalchemy>=2.0.0

# Web framework
flask>=2.3.0
flask-cors>=4.0.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0

# Utilities
schedule>=1.2.0
python-dateutil>=2.8.0
pytz>=2023.3

# Development and testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

